[SumitGPT - 001] better project structure and modulation
